+++ mldsa/native/x86_64/src/rej_uniform_eta2_avx2.c
@@ -60,7 +60,7 @@ unsigned int mld_rej_uniform_eta2_avx2(
 
     f1 = _mm256_sub_epi8(f0, bound);
     f0 = _mm256_sub_epi8(eta, f0);
-    good = _mm256_movemask_epi8(f1);
+    good = (uint32_t)_mm256_movemask_epi8(f1);
 
     g0 = _mm256_castsi256_si128(f0);
     g1 = _mm_loadl_epi64((__m128i *)&mld_rej_uniform_table[good & 0xFF]);
@@ -70,7 +70,7 @@ unsigned int mld_rej_uniform_eta2_avx2(
     f2 = _mm256_mullo_epi16(f2, p);
     f1 = _mm256_add_epi32(f1, f2);
     _mm256_storeu_si256((__m256i *)&r[ctr], f1);
-    ctr += _mm_popcnt_u32(good & 0xFF);
+    ctr += (unsigned)_mm_popcnt_u32(good & 0xFF);
     good >>= 8;
     pos += 4;
 
@@ -86,7 +86,7 @@ unsigned int mld_rej_uniform_eta2_avx2(
     f2 = _mm256_mullo_epi16(f2, p);
     f1 = _mm256_add_epi32(f1, f2);
     _mm256_storeu_si256((__m256i *)&r[ctr], f1);
-    ctr += _mm_popcnt_u32(good & 0xFF);
+    ctr += (unsigned)_mm_popcnt_u32(good & 0xFF);
     good >>= 8;
     pos += 4;
 
@@ -102,7 +102,7 @@ unsigned int mld_rej_uniform_eta2_avx2(
     f2 = _mm256_mullo_epi16(f2, p);
     f1 = _mm256_add_epi32(f1, f2);
     _mm256_storeu_si256((__m256i *)&r[ctr], f1);
-    ctr += _mm_popcnt_u32(good & 0xFF);
+    ctr += (unsigned)_mm_popcnt_u32(good & 0xFF);
     good >>= 8;
     pos += 4;
 
@@ -118,7 +118,7 @@ unsigned int mld_rej_uniform_eta2_avx2(
     f2 = _mm256_mullo_epi16(f2, p);
     f1 = _mm256_add_epi32(f1, f2);
     _mm256_storeu_si256((__m256i *)&r[ctr], f1);
-    ctr += _mm_popcnt_u32(good);
+    ctr += (unsigned)_mm_popcnt_u32(good);
     pos += 4;
   }
 
@@ -130,12 +130,12 @@ unsigned int mld_rej_uniform_eta2_avx2(
     if (t0 < 15)
     {
       t0 = t0 - (205 * t0 >> 10) * 5;
-      r[ctr++] = 2 - t0;
+      r[ctr++] = (int32_t)(2 - t0);
     }
     if (t1 < 15 && ctr < MLDSA_N)
     {
       t1 = t1 - (205 * t1 >> 10) * 5;
-      r[ctr++] = 2 - t1;
+      r[ctr++] = (int32_t)(2 - t1);
     }
   }
